Split/Move files verbatim into shared partition.
