+++ src/mscorlib/shared/System/Reflection/ObfuscateAssemblyAttribute.cs
