+++ src/mscorlib/shared/System/Reflection/ObfuscationAttribute.cs
