+++ src/mscorlib/shared/System/Reflection/TypeDelegator.cs
