+++ src/mscorlib/shared/System/Runtime/Serialization/SafeSerializationEventArgs.cs
@@ -229,13 +229,4 @@ namespace System.Runtime.Serialization
             get { return m_streamingContext; }
         }
     }
-
-    // Interface to be supported by objects which are stored in safe serialization stores
-    public interface ISafeSerializationData
-    {
-        // CompleteDeserialization is called when the object to which the extra serialized data was attached
-        // has completed its deserialization, and now needs to be populated with the extra data stored in
-        // this object.
-        void CompleteDeserialization(object deserialized);
-    }
 }
